api: Refactor, into own module
* Move the new Org API into it's own API submodule
* Make use of dataclass serializer library
