+++ datastore/api/experimental/serializers.py
@@ -1,137 +1,9 @@
 from rest_framework import serializers
-from rest_framework.reverse import reverse
-from drf_spectacular.utils import extend_schema_field
-from drf_spectacular.types import OpenApiTypes
 
 import db.models as db
-from . import models
 
 
-class GetterRunSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = db.GetterRun
-        fields = ["datetime", "archived"]
-
-
-class OrganisationRefSerializer(serializers.Serializer):
-    org_id = serializers.CharField()
-    self = serializers.SerializerMethodField()
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_self(self, org):
-        """Get the URL to this object's detail."""
-        return reverse(
-            "api:organisation-detail",
-            kwargs={"org_id": org.org_id},
-            request=self.context.get("request"),
-        )
-
-
-class GrantSerializer(serializers.ModelSerializer):
+class CurrentLatestGrantSerializer(serializers.ModelSerializer):
     class Meta:
         model = db.Grant
         exclude = ["id", "getter_run", "latest", "source_file"]
-
-    self = serializers.SerializerMethodField()
-    publisher = serializers.SerializerMethodField()
-    recipients = serializers.SerializerMethodField()
-    funders = serializers.SerializerMethodField()
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_self(self, grant):
-        """Get the URL to this object's detail."""
-        return reverse(
-            "api:grant-detail",
-            kwargs={"grant_id": grant.grant_id},
-            request=self.context.get("request"),
-        )
-
-    @extend_schema_field(OrganisationRefSerializer)
-    def get_publisher(self, grant):
-        return OrganisationRefSerializer(
-            models.OrganisationRef(grant.publisher.org_id), context=self.context
-        ).data
-
-    @extend_schema_field(OrganisationRefSerializer(many=True))
-    def get_recipients(self, grant):
-        return [
-            OrganisationRefSerializer(
-                models.OrganisationRef(recipient["id"]), context=self.context
-            ).data
-            for recipient in grant.data.get("recipientOrganization", [])
-        ]
-
-    @extend_schema_field(OrganisationRefSerializer(many=True))
-    def get_funders(self, grant):
-        return [
-            OrganisationRefSerializer(
-                models.OrganisationRef(funder["id"]), context=self.context
-            ).data
-            for funder in grant.data.get("fundingOrganization", [])
-        ]
-
-
-class OrganisationFunderSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = db.Funder
-        exclude = ["id"]
-
-
-class OrganisationRecipientSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = db.Recipient
-        exclude = ["id"]
-
-
-class OrganisationPublisherSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = db.Publisher
-        exclude = ["id", "getter_run"]
-
-
-class OrganisationListSerializer(serializers.Serializer):
-    self = serializers.SerializerMethodField()
-    org_id = serializers.CharField(max_length=200)
-    name = serializers.CharField(allow_blank=True, required=False)
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_self(self, org):
-        """Get the URL to this object's detail."""
-        request = self.context.get("request")
-        return reverse(
-            "api:organisation-detail", kwargs={"org_id": org["org_id"]}, request=request
-        )
-
-
-class OrganisationSerializer(serializers.Serializer):
-    self = serializers.SerializerMethodField()
-    org_id = serializers.CharField(max_length=200)
-    grants_made = serializers.SerializerMethodField()
-    grants_received = serializers.SerializerMethodField()
-    funder = OrganisationFunderSerializer(required=False)
-    recipient = OrganisationRecipientSerializer(required=False)
-    publisher = OrganisationPublisherSerializer(required=False)
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_self(self, org):
-        """Get the URL to this object's detail."""
-        request = self.context.get("request")
-        return reverse(
-            "api:organisation-detail", kwargs={"org_id": org.org_id}, request=request
-        )
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_grants_made(self, org):
-        return reverse(
-            "api:organisation-grants-made",
-            kwargs={"org_id": org.org_id},
-            request=self.context.get("request"),
-        )
-
-    @extend_schema_field(OpenApiTypes.URI)
-    def get_grants_received(self, org):
-        return reverse(
-            "api:organisation-grants-received",
-            kwargs={"org_id": org.org_id},
-            request=self.context.get("request"),
-        )
