+++ requirements.in
@@ -8,3 +8,4 @@ psycopg2==2.8
 prometheus_client==0.7
 django-environ
 drf-spectacular>=0.27,<0.28
+djangorestframework-dataclasses>=1.3.1,<2
