+++ requirements.txt
@@ -50,6 +50,7 @@ django==3.2.16
     #   django-filter
     #   django-prettyjson
     #   djangorestframework
+    #   djangorestframework-dataclasses
     #   drf-spectacular
 django-environ==0.9.0
     # via -r requirements.in
@@ -60,7 +61,10 @@ django-prettyjson==0.4.1
 djangorestframework==3.11.2
     # via
     #   -r requirements.in
+    #   djangorestframework-dataclasses
     #   drf-spectacular
+djangorestframework-dataclasses==1.3.1
+    # via -r requirements.in
 drf-spectacular==0.27.0
     # via -r requirements.in
 et-xmlfile==1.1.0
